put state in contract

+++ app/src/main/java/nz/co/plantandfood/mvvmtodoapp/presentation/screen/todo_list/TodoListContract.kt
@@ -12,7 +12,7 @@ class TodoListContract {
         data class ShowSnackbar(
             val message: String,
             val action: String? = null,
-            var duration: SnackbarDuration = SnackbarDuration.Short
+            var duration: SnackbarDuration = SnackbarDuration.Long
         ): Effect()
     }
